+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 1.1.0
+* Added ability to use builder in PiPMaterialApp.
+* Fixed a major bug where clicks near the boundary were not getting recognized in pip.
+* Added borderRadius and elevation parameters to PiPWidget.
+* Updated ReadMe.
+
 ## 1.0.0
 * Initial Release